Reproducing paper results - example code
Writing example scripts that should reproduce the systematic results
presented in the initial paper, obviously due to sampling and the
stochastic elements the results will have variations

Succesfully replicated job submission as well as statistical analysis
(not including the alpha approach). This includes hash values in: the
result files (.npz) the statistical analysis files (thermo), and the
analytic analysis results (analytic_results.json).

Next step is to finalize the plotting of the results:
 - Build out the flexible framework for simple plotting
 - Attempt to replicate the plots generated in the paper WITH the new
framework

If time permits it might be useful to have the ability to submit
plotting jobs to the server, although this is low priority.